pyproject.toml
[project]
name = "lambda-erp"
version = "0.1.0"
description = "Core ERP logic - accounting, sales, purchasing, inventory"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "openai>=1.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "python-dateutil>=2.8",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=5",
    "Jinja2>=3.1",
    "weasyprint>=62.0",
    "holidays>=0.40",
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
